Add validator for the application name (#37)
Before, when a user introduce an application name that includes spaces,
the script will exit, throwing an error.

In this PR, we confirm the application name by checking that:

- The name is just not numbers.
- The name doesn't include illegal characters. 
- If it includes spaces, we parse the name and return the camelized
name.

---------

+++ bin/belt.js
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 import { execSync } from 'child_process';
 import fs from 'fs-extra';
+import _ from 'lodash';
 import path from 'path';
 
 // executes the CLI locally from the `builds` directory
@@ -15,8 +16,9 @@ async function run() {
 
   // clean /builds, cd into it
   fs.mkdirSync(dir, { recursive: true });
-  const appName = process.argv[2];
-  if (fs.existsSync(path.join(dir, appName))) {
+  const rawAppName = process.argv[2];
+  const appName = _.upperFirst(_.camelCase((rawAppName || '').trim()));
+  if (appName && fs.existsSync(path.join(dir, appName))) {
     fs.rmSync(path.join(dir, appName), { recursive: true });
   }
   process.chdir(dir);
@@ -29,17 +31,19 @@ async function run() {
     },
   );
 
-  process.chdir(appName);
-  execSync(`npm run test:all`, { stdio: 'inherit' });
+  if (appName) {
+    process.chdir(appName);
+    execSync(`npm run test:all`, { stdio: 'inherit' });
 
-  process.chdir('../..');
-  console.log(`some commands you might want to run now:
+    process.chdir('../..');
+    console.log(`some commands you might want to run now:
 
   cd builds/${appName}
   cd builds/${appName} && npm run test:all
   cd builds/${appName} && npm run ios
   code builds/${appName}
   `);
+  }
 }
 
 function getNodeRunner() {
